requirements.txt
psycopg2
python-dotenv
pandas
psycopg2-binary
sqlalchemy